qcom: dma-heap: Update dma_heap_ops->allocate function signature
Commit 30d2da22a338 ("BACKPORT: FROMLIST: dma-buf/heaps: Correct the
types of fd_flags and heap_flags") modified the function signature
of dma_heap_ops->allocate. Update qcom implementation of this
callback.

+++ drivers/dma-buf/heaps/qcom_tvm_carveout_heap.c
@@ -278,8 +278,8 @@ static void tvm_heap_obj_release(struct qcom_sg_buffer *buffer)
 
 static struct dma_buf *tvm_heap_allocate(struct dma_heap *dma_heap,
 						unsigned long len,
-						unsigned long fd_flags,
-						unsigned long heap_flags)
+						u32 fd_flags,
+						u64 heap_flags)
 {
 	struct tvm_heap *heap = dma_heap_get_drvdata(dma_heap);
 	struct tvm_pool *pool;
